+++ tests/Encoder/TimestampLastCombCodecTest.php
@@ -19,7 +19,7 @@ class TimestampLastCombCodecTest extends TestCase
      */
     private $builderMock;
 
-    public function setUp()
+    protected function setUp()
     {
         $this->builderMock = $this->createMock('Ramsey\Uuid\Builder\UuidBuilderInterface');
         $this->codec = new TimestampLastCombCodec($this->builderMock);